operator: merge base and generic reconciler
