+++ integrations/operator/controllers/reconcilers/generic.go
@@ -21,7 +21,9 @@ package reconcilers
 import (
 	"context"
 	"fmt"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"reflect"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/gravitational/trace"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -33,6 +35,19 @@ import (
 	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/gravitational/teleport/api/types"
+)
+
+const (
+	// DeletionFinalizer is a name of finalizer added to Resource's 'finalizers' field
+	// for tracking deletion events.
+	DeletionFinalizer = "resources.teleport.dev/deletion"
+	// AnnotationFlagIgnore is the Kubernetes annotation containing the "ignore" flag.
+	// When set to true, the operator will not reconcile the CR.
+	AnnotationFlagIgnore = "teleport.dev/ignore"
+	// AnnotationFlagKeep is the Kubernetes annotation containing the "keep" flag.
+	// When set to true, the operator will not delete the Teleport Resource if the
+	// CR is deleted.
+	AnnotationFlagKeep = "teleport.dev/keep"
 )
 
 // Resource is any Teleport Resource the controller manages.
@@ -77,7 +92,7 @@ type resourceMutator[T Resource] interface {
 
 // resourceReconciler is a Teleport generic reconciler.
 type resourceReconciler[T any, K KubernetesCR[T]] struct {
-	ResourceBaseReconciler
+	kubeClient     kclient.Client
 	resourceClient resourceClient[T]
 	gvk            schema.GroupVersionKind
 	adapter        Adapter[T]
@@ -103,7 +118,7 @@ func (r resourceReconciler[T, K]) Upsert(ctx context.Context, obj kclient.Object
 	)
 	updateErr := updateStatus(updateStatusConfig{
 		ctx:         ctx,
-		client:      r.Client,
+		client:      r.kubeClient,
 		k8sResource: k8sResource,
 		condition:   getStructureConditionFromError(err),
 	})
@@ -118,7 +133,7 @@ func (r resourceReconciler[T, K]) Upsert(ctx context.Context, obj kclient.Object
 	existingResource, err := r.resourceClient.Get(ctx, name)
 	updateErr = updateStatus(updateStatusConfig{
 		ctx:         ctx,
-		client:      r.Client,
+		client:      r.kubeClient,
 		k8sResource: k8sResource,
 		condition:   getReconciliationConditionFromError(err, true /* ignoreNotFound */),
 	})
@@ -135,7 +150,7 @@ func (r resourceReconciler[T, K]) Upsert(ctx context.Context, obj kclient.Object
 		debugLog.Info("Resource is owned")
 		if updateErr = updateStatus(updateStatusConfig{
 			ctx:         ctx,
-			client:      r.Client,
+			client:      r.kubeClient,
 			k8sResource: k8sResource,
 			condition:   newOwnershipCondition,
 		}); updateErr != nil {
@@ -148,7 +163,7 @@ func (r resourceReconciler[T, K]) Upsert(ctx context.Context, obj kclient.Object
 		debugLog.Info("Resource does not exist yet")
 		if updateErr = updateStatus(updateStatusConfig{
 			ctx:         ctx,
-			client:      r.Client,
+			client:      r.kubeClient,
 			k8sResource: k8sResource,
 			condition:   newResourceCondition,
 		}); updateErr != nil {
@@ -172,7 +187,7 @@ func (r resourceReconciler[T, K]) Upsert(ctx context.Context, obj kclient.Object
 			// If an error happens we want to put it in status.conditions before returning.
 			updateErr = updateStatus(updateStatusConfig{
 				ctx:         ctx,
-				client:      r.Client,
+				client:      r.kubeClient,
 				k8sResource: k8sResource,
 				condition: metav1.Condition{
 					Type:    ConditionTypeSuccessfullyReconciled,
@@ -200,7 +215,7 @@ func (r resourceReconciler[T, K]) Upsert(ctx context.Context, obj kclient.Object
 	// If an error happens we want to put it in status.conditions before returning.
 	updateErr = updateStatus(updateStatusConfig{
 		ctx:         ctx,
-		client:      r.Client,
+		client:      r.kubeClient,
 		k8sResource: k8sResource,
 		condition:   getReconciliationConditionFromError(err, false /* ignoreNotFound */),
 	})
@@ -228,13 +243,91 @@ func (r resourceReconciler[T, K]) Delete(ctx context.Context, obj kclient.Object
 	return r.resourceClient.Delete(ctx, obj.GetName())
 }
 
-// Reconcile implements the controllers.Reconciler interface.
+// Reconcile receives an update request and reconcile the Resource,
+// it implements the controllers.Reconciler interface.
+//
+// When an event arrives we must propagate that change into the Teleport cluster.
+// We have two types of events: update/create and delete.
+//
+// For creating/updating we check if the Resource exists in Teleport
+// - if it does, we update it
+// - otherwise we create it
+// Always using the state of the Resource in the cluster as the source of truth.
+//
+// For deleting, the recommendation is to use finalizers.
+// Finalizers allow us to map an external Resource to a kubernetes Resource.
+// So, when we create or update a Resource, we add our own finalizer to the kubernetes Resource list of finalizers.
+//
+// For a delete event which has our finalizer: the Resource is deleted in Teleport.
+// If it doesn't have the finalizer, we do nothing.
+//
+// ----
+//
+// Every time we update a Resource in Kubernetes (adding finalizers or the OriginLabel), we end the reconciliation process.
+// Afterwards, we receive the request again and we progress to the next step.
+// This allow us to progress with smaller changes and avoid a long-running reconciliation.
+// */
 func (r resourceReconciler[T, K]) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	obj, err := GetUnstructuredObjectFromGVK(r.gvk)
 	if err != nil {
 		return ctrl.Result{}, trace.Wrap(err, "creating object in which the CR will be unmarshalled")
 	}
-	return r.Do(ctx, req, obj)
+	// https://sdk.operatorframework.io/docs/building-operators/golang/advanced-topics/#external-resources
+	log := ctrllog.FromContext(ctx).WithValues("namespacedname", req.NamespacedName)
+
+	if err := r.kubeClient.Get(ctx, req.NamespacedName, obj); err != nil {
+		if apierrors.IsNotFound(err) {
+			log.Info("not found")
+			return ctrl.Result{}, nil
+		}
+		log.Error(err, "failed to get Resource")
+		return ctrl.Result{}, trace.Wrap(err)
+	}
+
+	if isIgnored(obj) {
+		log.Info(fmt.Sprintf("Resource is flagged with annotation %q, it will not be reconciled.", AnnotationFlagIgnore))
+		return ctrl.Result{}, nil
+	}
+
+	hasDeletionFinalizer := controllerutil.ContainsFinalizer(obj, DeletionFinalizer)
+	isMarkedToBeDeleted := !obj.GetDeletionTimestamp().IsZero()
+
+	// Delete
+	if isMarkedToBeDeleted {
+		if hasDeletionFinalizer {
+			if isKept(obj) {
+				log.Info(fmt.Sprintf("Resource is flagged with annotation %q, it will not be deleted in Teleport.", AnnotationFlagKeep))
+			} else {
+				log.Info("deleting object in Teleport")
+				if err := r.Delete(ctx, obj); err != nil && !trace.IsNotFound(err) {
+					return ctrl.Result{}, trace.Wrap(err)
+				}
+			}
+
+			log.Info("removing finalizer")
+			controllerutil.RemoveFinalizer(obj, DeletionFinalizer)
+			if err := r.kubeClient.Update(ctx, obj); err != nil {
+				return ctrl.Result{}, trace.Wrap(err, "failed to remove finalizer after deleting in teleport")
+			}
+		}
+
+		// marked to be deleted without finalizer
+		return ctrl.Result{}, nil
+	}
+
+	if !hasDeletionFinalizer {
+		log.Info("adding finalizer")
+		controllerutil.AddFinalizer(obj, DeletionFinalizer)
+
+		err := r.kubeClient.Update(ctx, obj)
+
+		return ctrl.Result{}, trace.Wrap(err, "failed to add finalizer")
+	}
+
+	// Create or update
+	log.Info("upsert object in Teleport")
+	err = r.Upsert(ctx, obj)
+	return ctrl.Result{}, trace.Wrap(err)
 }
 
 // SetupWithManager implements the controllers.Reconciler interface.
@@ -254,6 +347,16 @@ func (r resourceReconciler[T, K]) SetupWithManager(mgr ctrl.Manager) error {
 			buildPredicate(),
 		).
 		Complete(r)
+}
+
+// isIgnored checks if the CR should be ignored
+func isIgnored(obj kclient.Object) bool {
+	return checkAnnotationFlag(obj, AnnotationFlagIgnore, false /* defaults to false */)
+}
+
+// isKept checks if the Teleport Resource should be kept if the CR is deleted
+func isKept(obj kclient.Object) bool {
+	return checkAnnotationFlag(obj, AnnotationFlagKeep, false /* defaults to false */)
 }
 
 // isResourceOriginKubernetes reads a teleport Resource metadata, searches for the origin label and checks its
