+++ integrations/operator/controllers/reconcilers/legacy_resource_with_labels.go
@@ -68,12 +68,10 @@ func NewTeleportResourceWithLabelsReconciler[T types.ResourceWithLabels, K Kuber
 		return nil, trace.Wrap(err)
 	}
 	reconciler := &resourceReconciler[T, K]{
-		ResourceBaseReconciler: ResourceBaseReconciler{Client: client},
-		resourceClient:         resourceClient,
-		gvk:                    gvk,
-		adapter:                ResourceWithLabelsAdapter[T]{},
+		kubeClient:     client,
+		resourceClient: resourceClient,
+		gvk:            gvk,
+		adapter:        ResourceWithLabelsAdapter[T]{},
 	}
-	reconciler.ResourceBaseReconciler.UpsertExternal = reconciler.Upsert
-	reconciler.ResourceBaseReconciler.DeleteExternal = reconciler.Delete
 	return reconciler, nil
 }
