+++ integrations/operator/controllers/reconcilers/legacy_resource_without_labels.go
@@ -82,12 +82,10 @@ func NewTeleportResourceWithoutLabelsReconciler[T resourceWithoutLabels, K Kuber
 		return nil, trace.Wrap(err)
 	}
 	reconciler := &resourceReconciler[T, K]{
-		ResourceBaseReconciler: ResourceBaseReconciler{Client: client},
-		resourceClient:         resourceClient,
-		gvk:                    gvk,
-		adapter:                ResourceWithoutLabelsAdapter[T]{},
+		kubeClient:     client,
+		resourceClient: resourceClient,
+		gvk:            gvk,
+		adapter:        ResourceWithoutLabelsAdapter[T]{},
 	}
-	reconciler.ResourceBaseReconciler.UpsertExternal = reconciler.Upsert
-	reconciler.ResourceBaseReconciler.DeleteExternal = reconciler.Delete
 	return reconciler, nil
 }
