+++ integrations/operator/controllers/reconcilers/resource153.go
@@ -76,12 +76,10 @@ func NewTeleportResource153Reconciler[T types.Resource153, K KubernetesCR[T]](
 		return nil, trace.Wrap(err)
 	}
 	reconciler := &resourceReconciler[T, K]{
-		ResourceBaseReconciler: ResourceBaseReconciler{Client: client},
-		resourceClient:         resourceClient,
-		gvk:                    gvk,
-		adapter:                Resource153Adapter[T]{},
+		kubeClient:     client,
+		resourceClient: resourceClient,
+		gvk:            gvk,
+		adapter:        Resource153Adapter[T]{},
 	}
-	reconciler.ResourceBaseReconciler.UpsertExternal = reconciler.Upsert
-	reconciler.ResourceBaseReconciler.DeleteExternal = reconciler.Delete
 	return reconciler, nil
 }
